first round of refactoring for extra constraints
