+++ torsiondrive/qm_engine.py
@@ -1,8 +1,6 @@
 import os
 import subprocess
 import numpy as np
-from collections import OrderedDict
-from geometric.nifty import uncommadash
 from geometric.molecule import Molecule
 from copy import deepcopy
 
@@ -12,76 +10,6 @@ def check_all_float(iterable):
         return True
     except ValueError:
         return False
-
-def constraint_match(cons_type, cons_spec, indices):
-    """ Check to see if a string-specified constraint matches a set of provided zero-based indices. """
-    s = cons_spec.split()
-    if cons_type in ['bond', 'distance']:
-        sMatch = [int(i)-1 for i in s[:2]]
-        if len(indices) != 2: return False
-        elif tuple(sMatch) == tuple(indices): return True
-        elif tuple(sMatch[::-1]) == tuple(indices): return True
-        else: return False
-    elif cons_type == 'angle':
-        sMatch = [int(i)-1 for i in s[:3]]
-        if len(indices) != 3: return False
-        elif tuple(sMatch) == tuple(indices): return True
-        elif tuple(sMatch[::-1]) == tuple(indices): return True
-        else: return False
-    elif cons_type == 'dihedral':
-        # Dihedrals should "match" if the two middle atoms are the same
-        sMatch = [int(i)-1 for i in s[:4]]
-        if len(indices) != 4: return False
-        elif (sMatch[1], sMatch[2]) == (indices[1], indices[2]): return True
-        elif (sMatch[1], sMatch[2]) == (indices[2], indices[1]): return True
-        else: return False
-    elif cons_type == 'xyz':
-        sMatch = uncommadash(cons_spec)
-        if sMatch == sorted(indices): return True
-        elif len(indices) == 4 and indices[0] in sMatch and indices[3] in sMatch: return True
-        else: return False
-    else:
-        raise RuntimeError("Problem detected with user-supplied extra constraints")
-        
-def make_constraints_dict(constraints_string, exclude=[]):
-    """ Create an ordered dictionary with constraints specification. 
-    
-    Parameters
-    ----------
-    constraints_string: str
-        String-formatted constraint specification consistent with geomeTRIC constraints.txt
-    exclude: list, optional
-        A list of zero-based constraint atom indices (e.g. dihedrals to be scanned over)
-        which should not be included in this dictionary
-    """  
-    constraints_mode = None
-    constraints_dict = OrderedDict([('freeze', []),('set', [])])
-    for line in constraints_string.split('\n'):
-        # Ignore anything after a comment
-        line = line.split('#')[0].lower().strip()
-        if len(line) == 0: continue
-        if line.startswith('$'):
-            if line == '$freeze':
-                constraints_mode = 'freeze'
-            elif line == '$set':
-                constraints_mode = 'set'
-            elif line == '$end':
-                constraints_mode = None
-            elif line == '$scan':
-                raise RuntimeError('Additional $scan constraints are not allowed')
-            else:
-                raise RuntimeError('Unrecognized token starting with $')
-        else:
-            if constraints_mode == None:
-                print('ERROR: ', line)
-                raise RuntimeError('Trying to read the above constraint line, but constraint mode is not set')
-            else:
-                spec_tuple = tuple(line.split())
-                if spec_tuple[0] not in ['bond','distance','angle','dihedral','xyz']:
-                    raise RuntimeError('Only bond, angle, and dihedral, xyz constraints are supported')
-                if any([constraint_match(spec_tuple[0], ' '.join(spec_tuple[1:]), excl) for excl in exclude]): continue
-                constraints_dict[constraints_mode].append(spec_tuple)
-    return constraints_dict
 
 class QMEngine(object):
     def __init__(self, input_file=None, work_queue=None, native_opt=False, extra_constraints=None):
@@ -113,14 +41,13 @@ class QMEngine(object):
 
     def write_constraints_txt(self):
         """ write a constraints.txt file for geomeTRIC """
+        constraints_string = ''
         if self.extra_constraints is None:
-            with open('constraints.txt', 'w') as outfile:
-                outfile.write("$set\n")
-                for d1, d2, d3, d4, v in self.dihedral_idx_values:
-                    # geomeTRIC use atomic index starting from 1
-                    outfile.write("dihedral %d %d %d %d %f\n" % (d1+1, d2+1, d3+1, d4+1, v))
+            constraints_string += "$set\n"
+            for d1, d2, d3, d4, v in self.dihedral_idx_values:
+                # geomeTRIC use atomic index starting from 1
+                constraints_string += f"dihedral {d1+1} {d2+1} {d3+1} {d4+1} {v}\n"
         else:
-            constraints_string = ''
             for key, value_list in self.extra_constraints.items():
                 if key == 'freeze':
                     if len(value_list) > 0:
@@ -135,8 +62,8 @@ class QMEngine(object):
                         constraints_string += ("dihedral %d %d %d %d %f\n" % (d1+1, d2+1, d3+1, d4+1, v))
                 else:
                     raise KeyError("constraints key %s is not recognized" % key)
-            with open('constraints.txt', 'w') as outfile:
-                outfile.write(constraints_string)             
+        with open('constraints.txt', 'w') as outfile:
+            outfile.write(constraints_string)
 
     def load_geomeTRIC_output(self):
         """ Load the optimized geometry and energy into a new molecule object and return """
@@ -534,7 +461,7 @@ class EngineTerachem(QMEngine):
             for line in self.tera_temp:
                 if line == "$!constraints@here":
                     if hasattr(self, 'constraintsStr'):
-                        # self.optblockStr will be set by self.optimize_native()
+                        # self.constraintsStr will be set by self.optimize_native()
                         terain.write(self.constraintsStr)
                 else:
                     terain.write(line)
